refactor: update import paths for user queries across multiple routes
