+++ scripts/perf/measure-generation.ts
@@ -18,8 +18,8 @@ import {
   type ProviderMetadata,
 } from '@/lib/ai/provider';
 import { client, db } from '@/lib/db/drizzle';
-import { createUser, getUserByClerkId } from '@/lib/db/queries';
 import { ATTEMPT_CAP } from '@/lib/db/queries/attempts';
+import { createUser, getUserByClerkId } from '@/lib/db/queries/users';
 import { generationAttempts, learningPlans, modules } from '@/lib/db/schema';
 
 interface Stats {
