+++ src/app/api/v1/plans/[planId]/attempts/route.ts
@@ -1,7 +1,8 @@
 import { withAuth, withErrorBoundary } from '@/lib/api/auth';
 import { NotFoundError, ValidationError } from '@/lib/api/errors';
 import { json } from '@/lib/api/response';
-import { getPlanAttemptsForUser, getUserByClerkId } from '@/lib/db/queries';
+import { getPlanAttemptsForUser } from '@/lib/db/queries/plans';
+import { getUserByClerkId } from '@/lib/db/queries/users';
 import { mapAttemptsToClient } from '@/lib/mappers/detailToClient';
 
 function getPlanId(req: Request) {
