+++ src/app/api/v1/plans/route.ts
@@ -7,7 +7,8 @@ import { AttemptCapExceededError, ValidationError } from '@/lib/api/errors';
 import { checkPlanGenerationRateLimit } from '@/lib/api/rate-limit';
 import { json } from '@/lib/api/response';
 import { db } from '@/lib/db/drizzle';
-import { getPlanSummariesForUser, getUserByClerkId } from '@/lib/db/queries';
+import { getPlanSummariesForUser } from '@/lib/db/queries/plans';
+import { getUserByClerkId } from '@/lib/db/queries/users';
 import { ATTEMPT_CAP } from '@/lib/db/queries/attempts';
 import { generationAttempts, learningPlans, modules } from '@/lib/db/schema';
 import { enqueueJob } from '@/lib/jobs/queue';
